Proposal to better keep tests in sync
As the code in test_set_state_noecho was mostly a copy, I tried to make it more DRY.
